MVS CPLD: don't go back to menu on RESET
There are a couple games on MVS (ie: mslug2) that have a check to try
and verify the MVS motherboard is an original one.  This check involves
writing some data to backup ram, and forcing the game to watchdog.  If
the watchdog doesn't trigger a reboot, the game will display a message
about using original MVS hardware.

A watchdog triggers a full reboot of the motherboard including a
low/high on the RESET line.  This was causing the MVS 161in1 cart to
go back to the menu if you tried to load one of these games.