Compiler: don't double up prom for <= 1MB games
For games that had <= 1MB prom the compiler was doubling them
up causing them to consume 2MB of output prom data.  Games of
this size should not be accessing the 1MB port region of memory,
since there is nothing there on the real cart.

This update keeps those games at 1MB in prom size.  If for some
reason the game tries to read from the port region, they will just
get mapped to the next games first 1MB.

Doing this saves 70MB+ and allows the full offical rom set + a
number of homebrew to fit within the 384MB prom flash size.